a little bit of reorg/cleanup
